Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-width-scroll.tentative-expected.txt


FAIL getComputedStyle() round-trips in presence of scrollbars assert_equals: width round-trips expected "85px" but got "70px"
